server: split code
