+++ server/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "stat.go",
         "statistics_handler.go",
         "tokenlimiter.go",
-        "util.go",
     ],
     importpath = "github.com/pingcap/tidb/server",
     visibility = ["//visibility:public"],
@@ -158,7 +157,6 @@ go_test(
         "tidb_library_test.go",
         "tidb_serial_test.go",
         "tidb_test.go",
-        "util_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":server"],
@@ -213,7 +211,6 @@ go_test(
         "//util/cpuprofile",
         "//util/dbterror/exeerrors",
         "//util/deadlockhistory",
-        "//util/mock",
         "//util/plancodec",
         "//util/replayer",
         "//util/resourcegrouptag",
