+++ server/internal/dump/dump.go
@@ -37,11 +37,22 @@ package dump
 
 import (
 	"encoding/binary"
+	"math"
+	"strconv"
 	"time"
 
+	"github.com/pingcap/tidb/errno"
+	"github.com/pingcap/tidb/parser/charset"
 	"github.com/pingcap/tidb/parser/mysql"
+	"github.com/pingcap/tidb/server/internal/column"
+	"github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/util/chunk"
+	"github.com/pingcap/tidb/util/dbterror"
+	"github.com/pingcap/tidb/util/hack"
 )
+
+var errInvalidType = dbterror.ClassServer.NewStd(errno.ErrInvalidType)
 
 // LengthEncodedString dumps a string as length encoded byte slice.
 func LengthEncodedString(buffer []byte, bytes []byte) []byte {
@@ -165,4 +176,135 @@ func BinaryDateTime(data []byte, t types.Time) []byte {
 		}
 	}
 	return data
+}
+
+func dumpTextRow(buffer []byte, columns []*column.Info, row chunk.Row, d *column.ResultEncoder) ([]byte, error) {
+	if d == nil {
+		d = column.NewResultEncoder(charset.CharsetUTF8MB4)
+	}
+	tmp := make([]byte, 0, 20)
+	for i, col := range columns {
+		if row.IsNull(i) {
+			buffer = append(buffer, 0xfb)
+			continue
+		}
+		switch col.Type {
+		case mysql.TypeTiny, mysql.TypeShort, mysql.TypeInt24, mysql.TypeLong:
+			tmp = strconv.AppendInt(tmp[:0], row.GetInt64(i), 10)
+			buffer = LengthEncodedString(buffer, tmp)
+		case mysql.TypeYear:
+			year := row.GetInt64(i)
+			tmp = tmp[:0]
+			if year == 0 {
+				tmp = append(tmp, '0', '0', '0', '0')
+			} else {
+				tmp = strconv.AppendInt(tmp, year, 10)
+			}
+			buffer = LengthEncodedString(buffer, tmp)
+		case mysql.TypeLonglong:
+			if mysql.HasUnsignedFlag(uint(columns[i].Flag)) {
+				tmp = strconv.AppendUint(tmp[:0], row.GetUint64(i), 10)
+			} else {
+				tmp = strconv.AppendInt(tmp[:0], row.GetInt64(i), 10)
+			}
+			buffer = LengthEncodedString(buffer, tmp)
+		case mysql.TypeFloat:
+			prec := -1
+			if columns[i].Decimal > 0 && int(col.Decimal) != mysql.NotFixedDec && col.Table == "" {
+				prec = int(col.Decimal)
+			}
+			tmp = util.AppendFormatFloat(tmp[:0], float64(row.GetFloat32(i)), prec, 32)
+			buffer = LengthEncodedString(buffer, tmp)
+		case mysql.TypeDouble:
+			prec := types.UnspecifiedLength
+			if col.Decimal > 0 && int(col.Decimal) != mysql.NotFixedDec && col.Table == "" {
+				prec = int(col.Decimal)
+			}
+			tmp = util.AppendFormatFloat(tmp[:0], row.GetFloat64(i), prec, 64)
+			buffer = LengthEncodedString(buffer, tmp)
+		case mysql.TypeNewDecimal:
+			buffer = LengthEncodedString(buffer, hack.Slice(row.GetMyDecimal(i).String()))
+		case mysql.TypeString, mysql.TypeVarString, mysql.TypeVarchar, mysql.TypeBit,
+			mysql.TypeTinyBlob, mysql.TypeMediumBlob, mysql.TypeLongBlob, mysql.TypeBlob:
+			d.UpdateDataEncoding(col.Charset)
+			buffer = LengthEncodedString(buffer, d.EncodeData(row.GetBytes(i)))
+		case mysql.TypeDate, mysql.TypeDatetime, mysql.TypeTimestamp:
+			buffer = LengthEncodedString(buffer, hack.Slice(row.GetTime(i).String()))
+		case mysql.TypeDuration:
+			dur := row.GetDuration(i, int(col.Decimal))
+			buffer = LengthEncodedString(buffer, hack.Slice(dur.String()))
+		case mysql.TypeEnum:
+			d.UpdateDataEncoding(col.Charset)
+			buffer = LengthEncodedString(buffer, d.EncodeData(hack.Slice(row.GetEnum(i).String())))
+		case mysql.TypeSet:
+			d.UpdateDataEncoding(col.Charset)
+			buffer = LengthEncodedString(buffer, d.EncodeData(hack.Slice(row.GetSet(i).String())))
+		case mysql.TypeJSON:
+			// The collation of JSON type is always binary.
+			// To compatible with MySQL, here we treat it as utf-8.
+			d.UpdateDataEncoding(mysql.DefaultCollationID)
+			buffer = LengthEncodedString(buffer, d.EncodeData(hack.Slice(row.GetJSON(i).String())))
+		default:
+			return nil, errInvalidType.GenWithStack("invalid type %v", columns[i].Type)
+		}
+	}
+	return buffer, nil
+}
+
+func dumpBinaryRow(buffer []byte, columns []*column.Info, row chunk.Row, d *column.ResultEncoder) ([]byte, error) {
+	if d == nil {
+		d = column.NewResultEncoder(charset.CharsetUTF8MB4)
+	}
+	buffer = append(buffer, mysql.OKHeader)
+	nullBitmapOff := len(buffer)
+	numBytes4Null := (len(columns) + 7 + 2) / 8
+	for i := 0; i < numBytes4Null; i++ {
+		buffer = append(buffer, 0)
+	}
+	for i := range columns {
+		if row.IsNull(i) {
+			bytePos := (i + 2) / 8
+			bitPos := byte((i + 2) % 8)
+			buffer[nullBitmapOff+bytePos] |= 1 << bitPos
+			continue
+		}
+		switch columns[i].Type {
+		case mysql.TypeTiny:
+			buffer = append(buffer, byte(row.GetInt64(i)))
+		case mysql.TypeShort, mysql.TypeYear:
+			buffer = Uint16(buffer, uint16(row.GetInt64(i)))
+		case mysql.TypeInt24, mysql.TypeLong:
+			buffer = Uint32(buffer, uint32(row.GetInt64(i)))
+		case mysql.TypeLonglong:
+			buffer = Uint64(buffer, row.GetUint64(i))
+		case mysql.TypeFloat:
+			buffer = Uint32(buffer, math.Float32bits(row.GetFloat32(i)))
+		case mysql.TypeDouble:
+			buffer = Uint64(buffer, math.Float64bits(row.GetFloat64(i)))
+		case mysql.TypeNewDecimal:
+			buffer = LengthEncodedString(buffer, hack.Slice(row.GetMyDecimal(i).String()))
+		case mysql.TypeString, mysql.TypeVarString, mysql.TypeVarchar, mysql.TypeBit,
+			mysql.TypeTinyBlob, mysql.TypeMediumBlob, mysql.TypeLongBlob, mysql.TypeBlob:
+			d.UpdateDataEncoding(columns[i].Charset)
+			buffer = LengthEncodedString(buffer, d.EncodeData(row.GetBytes(i)))
+		case mysql.TypeDate, mysql.TypeDatetime, mysql.TypeTimestamp:
+			buffer = BinaryDateTime(buffer, row.GetTime(i))
+		case mysql.TypeDuration:
+			buffer = append(buffer, BinaryTime(row.GetDuration(i, 0).Duration)...)
+		case mysql.TypeEnum:
+			d.UpdateDataEncoding(columns[i].Charset)
+			buffer = LengthEncodedString(buffer, d.EncodeData(hack.Slice(row.GetEnum(i).String())))
+		case mysql.TypeSet:
+			d.UpdateDataEncoding(columns[i].Charset)
+			buffer = LengthEncodedString(buffer, d.EncodeData(hack.Slice(row.GetSet(i).String())))
+		case mysql.TypeJSON:
+			// The collation of JSON type is always binary.
+			// To compatible with MySQL, here we treat it as utf-8.
+			d.UpdateDataEncoding(mysql.DefaultCollationID)
+			buffer = LengthEncodedString(buffer, d.EncodeData(hack.Slice(row.GetJSON(i).String())))
+		default:
+			return nil, errInvalidType.GenWithStack("invalid type %v", columns[i].Type)
+		}
+	}
+	return buffer, nil
 }
